Create individual specialism cols with specialist/generalist from specialisms_offered col
