+++ projects/_03_independent_cqc/_01_merge/jobs/merge_ind_cqc_data.py
@@ -47,6 +47,9 @@ cleaned_cqc_locations_columns_to_import = [
     CQCLClean.services_offered,
     CQCLClean.imputed_specialisms,
     CQCLClean.specialisms_offered,
+    CQCLClean.specialist_generalist_other_dementia,
+    CQCLClean.specialist_generalist_other_lda,
+    CQCLClean.specialist_generalist_other_mh,
     CQCLClean.related_location,
     CQCLClean.primary_service_type,
     CQCLClean.registered_manager_names,
