+++ utils/column_names/cleaned_data_files/cqc_location_cleaned.py
@@ -49,5 +49,8 @@ class CqcLocationCleanedColumns(NewCqcLocationApiColumns, ONSClean):
     )
     services_offered: str = "services_offered"
     specialisms_offered: str = "specialisms_offered"
+    specialist_generalist_other_dementia: str = "specialist_generalist_other_dementia"
+    specialist_generalist_other_lda: str = "specialist_generalist_other_lda"
+    specialist_generalist_other_mh: str = "specialist_generalist_other_mh"
     time_registered: str = "time_registered"
     time_since_dormant: str = "time_since_dormant"
