+++ utils/column_values/categorical_columns_by_dataset.py
@@ -168,6 +168,15 @@ class CleanedIndCQCCategoricalValues:
         IndCQC.ascwds_filtering_rule
     )
     related_location_column_values = RelatedLocation(CQCLClean.related_location)
+    specialist_generalist_other_dementia_column_values = SpecialistGeneralistOther(
+        IndCQC.specialist_generalist_other_dementia
+    )
+    specialist_generalist_other_lda_column_values = SpecialistGeneralistOther(
+        IndCQC.specialist_generalist_other_lda
+    )
+    specialist_generalist_other_mh_column_values = SpecialistGeneralistOther(
+        IndCQC.specialist_generalist_other_mh
+    )
 
 
 @dataclass
@@ -231,15 +240,6 @@ class EstimatedIndCQCFilledPostsCategoricalValues:
     )
     estimate_filled_posts_source_column_values = EstimateFilledPostsSource(
         IndCQC.estimate_filled_posts_source
-    )
-    specialist_generalist_other_dementia_column_values = SpecialistGeneralistOther(
-        IndCQC.specialist_generalist_other_dementia
-    )
-    specialist_generalist_other_lda_column_values = SpecialistGeneralistOther(
-        IndCQC.specialist_generalist_other_lda
-    )
-    specialist_generalist_other_mh_column_values = SpecialistGeneralistOther(
-        IndCQC.specialist_generalist_other_mh
     )
