+++ utils/validation/validation_rules/estimated_ind_cqc_filled_posts_validation_rules.py
@@ -28,9 +28,6 @@ class EstimatedIndCqcFilledPostsValidationRules:
             IndCqcColumns.unix_time,
             IndCqcColumns.estimate_filled_posts,
             IndCqcColumns.estimate_filled_posts_source,
-            IndCqcColumns.specialist_generalist_other_dementia,
-            IndCqcColumns.specialist_generalist_other_lda,
-            IndCqcColumns.specialist_generalist_other_mh,
         ],
         RuleName.index_columns: [
             IndCqcColumns.location_id,
@@ -73,9 +70,6 @@ class EstimatedIndCqcFilledPostsValidationRules:
             IndCqcColumns.current_region: CatValues.current_region_column_values.categorical_values,
             IndCqcColumns.ascwds_filled_posts_source: CatValues.ascwds_filled_posts_source_column_values.categorical_values,
             IndCqcColumns.estimate_filled_posts_source: CatValues.estimate_filled_posts_source_column_values.categorical_values,
-            IndCqcColumns.specialist_generalist_other_dementia: CatValues.specialist_generalist_other_dementia_column_values,
-            IndCqcColumns.specialist_generalist_other_lda: CatValues.specialist_generalist_other_lda_column_values,
-            IndCqcColumns.specialist_generalist_other_mh: CatValues.specialist_generalist_other_mh_column_values,
         },
         RuleName.distinct_values: {
             IndCqcColumns.care_home: CatValues.care_home_column_values.count_of_categorical_values,
@@ -85,9 +79,6 @@ class EstimatedIndCqcFilledPostsValidationRules:
             IndCqcColumns.current_region: CatValues.current_region_column_values.count_of_categorical_values,
             IndCqcColumns.ascwds_filled_posts_source: CatValues.ascwds_filled_posts_source_column_values.count_of_categorical_values,
             IndCqcColumns.estimate_filled_posts_source: CatValues.estimate_filled_posts_source_column_values.count_of_categorical_values,
-            IndCqcColumns.specialist_generalist_other_dementia: CatValues.specialist_generalist_other_dementia_column_values.count_of_categorical_values,
-            IndCqcColumns.specialist_generalist_other_lda: CatValues.specialist_generalist_other_lda_column_values.count_of_categorical_values,
-            IndCqcColumns.specialist_generalist_other_mh: CatValues.specialist_generalist_other_mh_column_values.count_of_categorical_values,
         },
         RuleName.custom_type: CustomValidationRules.care_home_and_primary_service_type,
         # RuleName.custom_type: CustomValidationRules.cqc_gac_services_and_primary_service_type_second_level,
